requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
certifi==2020.6.20
dotty-dict==1.2.1
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
jsonschema==3.2.0
numpy==1.19.1
pandas==1.1.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.2
requests==2.24.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
rtstatlib==1.0.0
setuptools-scm==4.1.2
six==1.15.0
typing-extensions==3.7.4.3
urllib3
zipp==3.1.0
flywheel-bids==0.9.1
flywheel-gear-toolkit==0.1.4
flywheel-gears==0.2.0
flywheel-sdk==14.4.0